pyproject.toml
[tool.poetry]
name = "template-paper"
version = "0.0.0"
description = ""
authors = ["Gael Close <gcl@melexis.com>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
numpy = "^1.23.3"
pandas = "^1.5.0"
python-box = "^6.0.2"
mergedeep = "^1.3.4"
tabulate = "^0.9.0"
forallpeople = "^2.6.3"
si-prefix = "^1.2.2"
runfile = "^1.0.7"
jupyter = "^1.0.0"
ipykernel = "^6.16.0"
jupytext = "^1.14.1"
pyprojroot = "0.2.0"
openpyxl = "^3.0.10"
pytest = "^7.1.3"
jupyter-cache = "^0.5.0"
GitPython = "^3.1.29"
pytest-plt = "^1.1.0"
jupyter-contrib-nbextensions = "^0.7.0"
jupyter-nbextensions-configurator = "^0.6.3"
seaborn = "^0.13.0"
matplotlib = "3.8.0"
Pint = "^0.22"
Pint-Pandas = "^0.5"
sigfig = "^1.3.3"
hvplot = "^0.9.0"
watermark = "^2.4.3"

